Reorganize API files
